Color Detection affecting game
Made it so color detection affects gameplay. When red is seen Bomb Prefab will spawn in in front of the player. When green is detected the speed power up with spawn in front of the player. Also added a cooldown period between when colors can be detected so then it will not make the game lag.